requirements.txt
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.1.0
pytest==7.4.3
pytest-mock==3.12.0
